Support separate upstream & peer MaxBatchSize
For high volume deployments of Refinery, a large batch size is necessary
to prevent the peer queue from overflowing. However, the large batch
sizes that are necessary to fix that issue also cause rejections from
the upstream API due to too large of payloads.

This commit enabled separate config for the Upstream and Peer batch
sizes.

Fixes #381.